theming manger development +minmal documentation
